scripts/requirements.txt
httpx>=0.27.0
pycryptodome>=3.20.0
